feat(bgatlas): observer ranking stats by visited cells
