+++ client/scripts/atlas/controllers/AtlasLegendController.js
@@ -1,5 +1,5 @@
 var utils = require('../utils')
 
-module.exports = function AtlasLegendController () {
+module.exports = /* @ngInject */function AtlasLegendController () {
   this.colors = utils.colors
 }
