+++ client/views/atlas/_stats_ranking.html
@@ -1,0 +1,18 @@
+<div class="page-atlas-stats-cell">
+  <div ng-if="$ctrl.loading" class="text-center">
+    <i class="fa fa-spinner fa-spin fa-3x"></i>
+  </div>
+
+  <div ng-if="!$ctrl.loading" class="table-responsive">
+    <table class="table table-striped table-hover table-condensed">
+      <tr ng-repeat="record in ::$ctrl.rows"
+          class="{{ ::record.user.id === $root.$user.getIdentity().id ? 'warning':'' }}">
+        <td class="text-right" style="width: 1px;">{{::record.position}}.</td>
+        <td>{{::record.user | userName}}</td>
+        <td class="text-right" translate translate-values="{count: record.count}">
+          ATLAS_STATS_RANKING_LABEL_COUNT
+        </td>
+      </tr>
+    </table>
+  </div>
+</div>
